Lower the padding on narrower viewports.
Added a media query to the css file to lower the left and right padding
when the width of the viewport is lower than 750px. This makes browsing
n-o-d-e.net, especially watching embedded videos in small windows or on
mobile phones better, since there isn't so much wasted space on either
side.